requirements.txt
# Azure SDK packages
azure-identity==1.15.0
azure-mgmt-videoanalyzer==1.0.0b1
azure-media-videoanalyzeredge==1.0.0b1
azure-ai-videoindexer==1.0.0b1
azure-storage-blob==12.19.0
azure-mgmt-cdn==12.0.0
azure-mgmt-frontdoor==1.0.0
azure-synapse-artifacts==0.17.0
azure-synapse-spark==0.7.0

# Web framework
fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
requests==2.31.0
httpx==0.27.0

# Database
pyodbc==5.0.1
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.22
Pillow==10.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
